+++ src/IntervalScheduling.cpp
@@ -1,7 +1,7 @@
 #include "IntervalScheduling.hpp"
 #include <algorithm>
 
-WeightedInterval::WeightedInterval(int s, int e, float sc) {
+WeightedInterval::WeightedInterval(int s, int e, double sc) {
     if (s < e) {
         start = s;
         end = e;
